Extract scaffold support class
